strlen strchr strrchr

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: jvictor- <jvictor-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/07 22:48:57 by jvictor-          #+#    #+#             */
-/*   Updated: 2021/06/07 23:15:55 by jvictor-         ###   ########.fr       */
+/*   Updated: 2021/06/12 12:07:57 by jvictor-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int		i;
-	int		len;
+	size_t		i;
+	size_t		len;
 
 	i = 0;
 	len = ft_strlen((char *)s);
