+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: jvictor- <jvictor-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/07 23:25:50 by jvictor-          #+#    #+#             */
-/*   Updated: 2021/06/07 23:58:00 by jvictor-         ###   ########.fr       */
+/*   Updated: 2021/06/12 12:09:12 by jvictor-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int		i;
-	int		len;
+	size_t		i;
+	size_t		len;
 
 	len = ft_strlen((char *)s);
 	i = len - 1;
